requirements.txt
langfuse>=3.0.0
langchain>=0.1.0
langchain-core>=0.1.0
langchain-google-genai>=0.0.5
google-generativeai>=0.3.0
python-dotenv>=1.0.0
scikit-learn>=1.0.0 